Title: 중앙값 계산 2, Time: 7ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241230/중앙값 계산 2/get-median-2.cpp
@@ -18,7 +18,7 @@ int main() {
 
     sort(arr, arr+n);
 
-   int m=(arr[0]+arr[n-1])/2;
+   float m=(arr[0]+arr[n-1])/2;
     int cnt=0;
     for(int i=0; i<n; i++)
     {
